on click sur carte avec ouverture de modale
