+++ src/style/favorites.scss
@@ -34,4 +34,14 @@
     padding-left: 20px;
     background-color: red;
   }
+
+
+
+  .open-modal{
+
+    width: 170px;
+    height: 230px;
+    
+
+  }
 }
